Finished KNN classification
Estimated run-time for grid search is 20min.
Came up to optimal hyper-parameters with leaf_size=1, n_neighbors=18 and p=1.
The score of training data is 1, which is a bit bad/overfitting? The scores for test data and validate data are both around 82%, which is good enough.
Without grid search, they are 87%, 80.4% and 80.3% respectively. So, we see a slightly improvement(around 1% only)